backend/app/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
python-dotenv